requirements.txt
matplotlib==3.7.1
PyYAML==6.0.1
PyYAML==6.0.1
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
